Fix integer signedness comparison warning in parameters.cpp
Removed the explicit `(int)` cast from `past_results[0]->utilization.size()`
in `utilization_power_valid(const std::string &u)` to properly compare
the `unsigned int index` against the `size_t` return type without warnings.

+++ src/parameters/parameters.cpp
@@ -404,7 +404,7 @@ int utilization_power_valid(const std::string &u)
 	if (past_results.size() == 0)
 		return 0;
 
-	if (index >= (int)past_results[0]->utilization.size())
+	if (index >= past_results[0]->utilization.size())
 		return 0;
 
 	first_value = past_results[0]->utilization[index];
